refactor(core): shared code optimization
shared code optimization
